Root Choices - Framework Started
Instead of just running the "C" choices while loop at start of program, it will now ask their intial choice "a, b or c". At this point when chosen "c" it will take the user through the C branch choices which is essentially a linear While Loop.

This starts the framework for the user to intiate the other branches such as "A" and "B"