Finished C Branch
C Branch is now implemented. User can go from c -> c1 - > c2 -> c3 in it's entirety. If an invalid input is made at any step, then it will quit the program. Also, by some quick testing I can see that it's not possible to input a previous charecter that's already in the UserChoice list, which will then jump it back to that function again. This was a problem with Prototype 1 where you could keep looping back to previous parts of the branch and never end the program.