Node definitions... Duplicacy problems
The idea was to create global variables names after each Node that will appear in the function. Then, when that Node is avaliable to be chosen by user, it's definition would then be printed. Issue is this; This idea was to reduce code/definition duplication to make the codebase cleaner. But everytime I printed a global variable it would then print it's position within the memory address. So, I've had to state that it's a "global" within the function and then reassign it's definition value. This will then pritn what i want it to print, but this is duplicating the code and makes the existence of having these Nodes as Global Variables rather pointless. A new way to make this work needs to be found because once the User has made all the choices that they can make, I want the program to then print all their choices within sequential order.