Quest Branch - Act 2 - Part 1
Creating the first "fractal" within in Act 2. Gives the user the choice once they reached Node "Q16" that they can go straight to Node Q34 (Which starts Act 3) or can go to Node Q17, which reiterates through the decision options again. The user will be able to reiterate at a maximum of 3 times.