Quest Plot - Act 2 - Part 2 - Issues fixed
There was an issue where if the User had previously chosen Node "q5" then when it came to Node Q8 or Q17, instead of giving them a choice to choose their "Buddy" quest Nodes, it would automatically take them there. This is a problem because it then neglects the Protagonist's own events/quest Nodes.

Fractal part 2 of 3 was implelmented too.