Adventure Branch - Print All Node Descriptor
Found a way to make it so the user is informed in the console on what their node option choices actually are. then ti stores that descriptive data and puts it in it's own unique list called "TemplateList".

Once the User has gotten to the end of one of the code branches, it then prints out each element within that list and gives a "template" on what the user has chosen.